Add auth for mongodb, remove not working errand
This commit removes the errand create-users since it’s not working
(the implementation is simply not finished).

Also this adds support for auth to the mongodb3 job.

New variables:

mongodb.auth.enable: If auth is enabled (Default: false)

Whenver this value is set to true, the mongod is starded with
the flag `--auth`.

To get this up and running the following steps are required:

1. Keep mongodb.auth.enable disabled (read: set it to false)
2. Deploy mongodb
3. Run errand replset to create the cluster
4. Run errand create-user to create an user to auth
5. Change mongodb.auth.enable to true
6. Update the deployment